Refactoring: ui.render is now subpackage of ui.schematics
