+++ src/main/java/de/uni_hannover/sra/minimax_simulator/model/machine/base/display/MachineDisplay.java
@@ -1,7 +1,7 @@
 package de.uni_hannover.sra.minimax_simulator.model.machine.base.display;
 
 import de.uni_hannover.sra.minimax_simulator.ui.layout.Dimension;
-import de.uni_hannover.sra.minimax_simulator.ui.render.RenderEnvironment;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.RenderEnvironment;
 import de.uni_hannover.sra.minimax_simulator.ui.schematics.SpriteOwner;
 
 import java.util.Set;
