+++ src/main/java/de/uni_hannover/sra/minimax_simulator/model/machine/minimax/MinimaxDisplay.java
@@ -6,8 +6,8 @@ import de.uni_hannover.sra.minimax_simulator.model.machine.base.display.FontMetr
 import de.uni_hannover.sra.minimax_simulator.model.machine.base.display.MachineDisplay;
 import de.uni_hannover.sra.minimax_simulator.model.machine.base.display.MachineDisplayListener;
 import de.uni_hannover.sra.minimax_simulator.model.machine.minimax.group.Group;
-import de.uni_hannover.sra.minimax_simulator.ui.render.DummyRenderEnvironment;
-import de.uni_hannover.sra.minimax_simulator.ui.render.RenderEnvironment;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.DummyRenderEnvironment;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.RenderEnvironment;
 import de.uni_hannover.sra.minimax_simulator.ui.schematics.SpriteOwner;
 import javafx.scene.text.Font;
