+++ src/main/java/de/uni_hannover/sra/minimax_simulator/model/machine/part/CuLabel.java
@@ -1,6 +1,6 @@
 package de.uni_hannover.sra.minimax_simulator.model.machine.part;
 
-import de.uni_hannover.sra.minimax_simulator.ui.render.Sprite;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.Sprite;
 import de.uni_hannover.sra.minimax_simulator.ui.schematics.parts.CuLabelSprite;
 
 /**
