--- src/main/java/de/uni_hannover/sra/minimax_simulator/ui/render/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Contains all classes for rendering.
- */
-// TODO: this is used by the ui.schematics package; move there?
-package de.uni_hannover.sra.minimax_simulator.ui.render;
