+++ src/main/java/de/uni_hannover/sra/minimax_simulator/ui/schematics/DefaultSpriteFactory.java
@@ -1,7 +1,7 @@
 package de.uni_hannover.sra.minimax_simulator.ui.schematics;
 
-import de.uni_hannover.sra.minimax_simulator.ui.render.Sprite;
-import de.uni_hannover.sra.minimax_simulator.ui.render.SpriteFactory;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.Sprite;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.SpriteFactory;
 
 /**
  * Default {@link SpriteFactory} with sprite owner class {@link SpriteOwner}.
