+++ src/main/java/de/uni_hannover/sra/minimax_simulator/ui/schematics/MachineSchematics.java
@@ -3,8 +3,8 @@ package de.uni_hannover.sra.minimax_simulator.ui.schematics;
 import de.uni_hannover.sra.minimax_simulator.ui.layout.Dimension;
 import de.uni_hannover.sra.minimax_simulator.model.machine.base.Machine;
 import de.uni_hannover.sra.minimax_simulator.model.machine.base.display.MachineDisplayListener;
-import de.uni_hannover.sra.minimax_simulator.ui.render.DefaultRenderEnvironment;
-import de.uni_hannover.sra.minimax_simulator.ui.render.SpriteCanvas;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.DefaultRenderEnvironment;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.SpriteCanvas;
 import javafx.scene.text.Font;
 
 import static com.google.common.base.Preconditions.checkNotNull;
