+++ src/main/java/de/uni_hannover/sra/minimax_simulator/ui/schematics/SpriteOwner.java
@@ -1,6 +1,6 @@
 package de.uni_hannover.sra.minimax_simulator.ui.schematics;
 
-import de.uni_hannover.sra.minimax_simulator.ui.render.Sprite;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.Sprite;
 
 /**
  * A {code SpriteOwner} is a class that contains {@link Sprite}s.
