+++ src/main/java/de/uni_hannover/sra/minimax_simulator/ui/schematics/parts/CircuitSprite.java
@@ -4,8 +4,8 @@ import de.uni_hannover.sra.minimax_simulator.Config;
 import de.uni_hannover.sra.minimax_simulator.ui.layout.Bounds;
 import de.uni_hannover.sra.minimax_simulator.ui.layout.Point;
 import de.uni_hannover.sra.minimax_simulator.model.machine.part.Pin;
-import de.uni_hannover.sra.minimax_simulator.ui.render.RenderEnvironment;
-import de.uni_hannover.sra.minimax_simulator.ui.render.Sprite;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.RenderEnvironment;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.Sprite;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Color;
 import javafx.scene.transform.Affine;
