+++ src/main/java/de/uni_hannover/sra/minimax_simulator/ui/schematics/parts/ConstantSprite.java
@@ -2,7 +2,6 @@ package de.uni_hannover.sra.minimax_simulator.ui.schematics.parts;
 
 import de.uni_hannover.sra.minimax_simulator.ui.layout.Bounds;
 import de.uni_hannover.sra.minimax_simulator.model.machine.part.Constant;
-import de.uni_hannover.sra.minimax_simulator.ui.render.RenderEnvironment;
 import javafx.scene.canvas.GraphicsContext;
 
 import static com.google.common.base.Preconditions.checkNotNull;
