+++ src/main/java/de/uni_hannover/sra/minimax_simulator/ui/schematics/parts/JunctionSprite.java
@@ -1,8 +1,8 @@
 package de.uni_hannover.sra.minimax_simulator.ui.schematics.parts;
 
 import de.uni_hannover.sra.minimax_simulator.model.machine.part.Junction;
-import de.uni_hannover.sra.minimax_simulator.ui.render.RenderEnvironment;
-import de.uni_hannover.sra.minimax_simulator.ui.render.Sprite;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.RenderEnvironment;
+import de.uni_hannover.sra.minimax_simulator.ui.schematics.render.Sprite;
 import javafx.scene.canvas.GraphicsContext;
 
 import static com.google.common.base.Preconditions.checkNotNull;
