+++ src/main/java/de/uni_hannover/sra/minimax_simulator/ui/schematics/render/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains all classes for rendering.
+ */
+package de.uni_hannover.sra.minimax_simulator.ui.schematics.render;
